Add initial link.py tests Optimize and tweak dedup tests PEP8 compliance on link.py and link_base.py Add truncated deduped test input files

+++ lib/cdi-linking/cdilinker/test/test_linker_dedup.py
@@ -3,6 +3,7 @@ import os
 import pandas as pd
 import pytest
 
+from cdilinker.linker.files import LinkFiles
 from cdilinker.linker.dedup import DeDeupProject
 
 
@@ -18,9 +19,14 @@ class TestLinkerDedup(object):
         yield project
 
         # Teardown and clean up
-        os.remove(project['output_root'] + 'dedup_matched.csv')
-        os.remove(project['output_root'] + 'deduped_data.csv')
-        os.remove(project['output_root'] + project['name'] + "_summary.pdf")
+        if os.path.isfile(project['output_root'] + 'dedup_matched.csv'):
+            os.remove(project['output_root'] + 'dedup_matched.csv')
+        if os.path.isfile(project['output_root'] + 'deduped_data.csv'):
+            os.remove(project['output_root'] + 'deduped_data.csv')
+        if os.path.isfile(project['output_root'] + project['name'] +
+                          '_summary.pdf'):
+            os.remove(project['output_root'] + project['name'] +
+                      '_summary.pdf')
 
     def test_init_none(self):
         """Ensure initialization does not proceed with empty JSON"""
@@ -57,6 +63,7 @@ class TestLinkerDedup(object):
         ddp.load_data()
 
         assert ddp.left_columns is not None
+        assert len(ddp.left_columns) == 6
         assert ddp.left_dtypes is not None
         assert ddp.left_dataset is not None
 
@@ -66,7 +73,8 @@ class TestLinkerDedup(object):
         ddp.load_data()
         ddp.run()
 
-        assert not os.path.isfile(project['output_root'] + 'matched_temp.csv')
+        assert not os.path.isfile(project['output_root'] +
+                                  LinkFiles.TEMP_MATCHED_FILE)
         assert ddp.steps is not None
         assert len(ddp.steps) == len(project['steps'])
         assert ddp.linked is not None
@@ -84,4 +92,4 @@ class TestLinkerDedup(object):
         assert ddp.total_entities == 98
         assert os.path.isfile(project['output_root'] + 'deduped_data.csv')
         assert os.path.isfile(project['output_root'] + project['name'] +
-                              "_summary.pdf")
+                              '_summary.pdf')
